fix: moved httpRequest function from utils.ts to services
