+++ src/__tests__/unit/analytics.spec.ts
@@ -1,12 +1,15 @@
 import { basename } from 'path';
 import { shareAnalyticsWithAppScript } from '../../analytics';
-import { getAppScriptUrl, getJobId, httpRequest } from '../../utils';
+import { getAppScriptUrl, getJobId } from '../../utils';
+import { httpRequest } from '../../services/http.service';
 import logger from '../../../logger';
 import { filePath, fileName, fileContent } from '../../types/renameFiles';
 
 jest.mock('../../utils', () => ({
     getAppScriptUrl: jest.fn(),
     getJobId: jest.fn(),
+}));
+jest.mock('../../services/http.service', () => ({
     httpRequest: jest.fn(),
 }));
